functions/example/requirements.txt
boto3>=1.26.133
python-decouple>=3.8
python-dotenv>=0.19
